app/requirements.txt
Flask==3.1.0
pymongo==4.11.2
argon2-cffi==23.1.0
PyJWT==2.10.0
APScheduler==3.11.0
python-gnupg==0.5.4
gunicorn==21.2.0
